--- src/db.py
@@ -1,61 +1,0 @@
-#!/usr/bin/env python3
-from flask_sqlalchemy import SQLAlchemy
-from schema_validation import perm_routes
-
-db = SQLAlchemy()
-
-class Client(db.Model):
-    __tablename__ = "klienti"
-
-    id      = db.Column("klient_id", db.Integer, primary_key=True)
-    name    = db.Column("knazov", db.String, nullable=False)
-    ico     = db.Column("kico", db.String, nullable=False)
-    zruseny = db.Column(db.String, nullable=False, default=0)
-
-    contacts = db.relationship("Contact", back_populates="client")
-
-    def as_dict(self):
-        columns = ["id", "name", "ico"] # Columns that should be included in the conversion
-        return {name: getattr(self, name) for name in columns}
-
-class Contact(db.Model):
-    __tablename__ = "klient_kontakt"
-
-    id           = db.Column("kkontakt_id", db.Integer, primary_key=True)
-    client_id    = db.Column("kklient_id", db.Integer, db.ForeignKey("klienti.klient_id"))
-    name         = db.Column("kkmeno", db.String, nullable=False)
-    phone_number = db.Column("kktel", db.String)
-    email        = db.Column("kkemail", db.String)
-
-    client = db.relationship("Client", back_populates="contacts")
-
-    def as_dict(self):
-        columns = ["id", "client_id", "name", "phone_number", "email"] # Columns that should be included in the conversion
-        return {name: getattr(self, name) for name in columns}
-
-class ApiKey(db.Model):
-    __tablename__ = "api_kluce"
-
-    id                    = db.Column("id", db.Integer, primary_key=True)
-    key                   = db.Column("kluc", db.String, unique=True)
-
-    perms = db.relationship("ApiPermission", back_populates="key", cascade="all,delete-orphan")
-
-    def as_dict(self):
-        # This is used to fill all routes that this key doesn't have permissions for with 0
-        zero_perms = {route: 0 for route in perm_routes}
-        actual_perms = {perm.route: perm.level for perm in self.perms}
-        return {
-            "id": self.id,
-            "permissions": {**zero_perms, **actual_perms}
-        }
-
-class ApiPermission(db.Model):
-    __tablename__ = "api_prava"
-
-    id = db.Column("id", db.Integer, primary_key=True)
-    key_id = db.Column("kluc_id", db.Integer, db.ForeignKey("api_kluce.id"))
-    route = db.Column("route", db.String, nullable=False)
-    level = db.Column("level", db.Integer, nullable=False)
-
-    key = db.relationship("ApiKey", back_populates="perms")
